calibrate-images/requirements.txt
Flask
google-cloud-firestore
google-cloud-storage
ruamel.yaml
panoptes-utils[images]>=0.2.32
